+++ server/src/main/resources/db/migration/V1__Implement_db_schema.sql
@@ -31,9 +31,9 @@ CREATE TABLE assets
 
 CREATE TABLE prices
 (
-    id           SERIAL PRIMARY KEY,
+    id           BIGSERIAL PRIMARY KEY,
     asset_id     INTEGER                  NOT NULL,
-    price        INTEGER                  NOT NULL,
+    price        BIGINT                   NOT NULL,
     set_datetime TIMESTAMP WITH TIME ZONE NOT NULL,
     FOREIGN KEY (asset_id) REFERENCES assets (id) ON DELETE CASCADE
 );
@@ -61,7 +61,7 @@ CREATE TABLE diversifications
     user_id         INTEGER                  NOT NULL,
     create_datetime TIMESTAMP WITH TIME ZONE NOT NULL,
     risk_type_id    INTEGER                  NOT NULL,
-    amount          INTEGER                  NOT NULL,
+    amount          BIGINT                   NOT NULL,
     FOREIGN KEY (user_id) REFERENCES users (id) ON DELETE SET NULL
 );
 
@@ -70,7 +70,7 @@ CREATE TABLE diversifications_assets
     id                 SERIAL PRIMARY KEY,
     diversification_id INTEGER NOT NULL,
     asset_id           INTEGER NOT NULL,
-    count              INTEGER NOT NULL,
+    count              BIGINT  NOT NULL,
     FOREIGN KEY (diversification_id) REFERENCES diversifications (id) ON DELETE SET NULL,
     FOREIGN KEY (asset_id) REFERENCES assets (id) ON DELETE SET NULL
 );
